pyproject.toml
[project]
name = "my-clip-image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.2",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "uvicorn[standard]>=0.38.0",
]
